api/requirements.txt
fastapi
starlette
tensorflow
numpy
opencv-python-headless
redis
pillow
